backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.15
sentence-transformers==2.3.1
huggingface-hub==0.19.4
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
slack-sdk==3.26.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
httpx==0.25.2